Kernel: Do not change GPIO settings for the RPI SD Host Controller
Different models of the Raspberry Pi have their microSD slot connected
to different GPIOs, and this information is only documented in the
device tree. Luckily, the firmware passes control to us with everything
configured in the right way, so we don't need to worry about it.

This fixes writing to the SD card on bare metal.