LibJS: Create const variables in ForIn/OfBodyEvaluation in strict mode
Our implementation of environment.CreateImmutableBinding(name, true)
in this AO was not correctly initializing const variables in strict
mode. This would mean that constant declarations in for loop bodies
would not throw if they were modified.

To fix this, add a new parameter to CreateVariable to force strict mode.

This fixes two of our test-js tests, no change to test262.

Co-Authored-By: Daniel Bertalan <dani@danielbertalan.dev>